third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-load-expected.txt

This is a testharness.js-based test.
FAIL invoking resource selection with load() assert_equals: networkState in separate script expected 0 but got 3
Harness: the test ran to completion.
